factored scrubbing and tokenization out of embed.py
